requirements.txt
bidict==0.21.4
dnspython==1.16.0
eventlet==0.30.2
greenlet==1.1.2
gunicorn==20.1.0
netifaces==0.10.6
python-engineio==4.3.0
python-socketio==5.4.1
six==1.16.0
